Implementation of `search' spec is rearranged.
* twittering-mode.el (twittering-timeline-spec-to-host-method):
return query string as a element of returned list.
(twittering-host-method-to-timeline-spec): receive query string as
a optional argument.
(twittering-http-get): remove special adjustment for search
method, which is now unnecessary.
(twittering-get-tweets): always receive query string as an
argument for search method.
(twittering-get-search): removed.
(twittering-get-and-render-timeline): send query string if the
given spec is `search'.
(twittering-search): execute search via
`twittering-get-and-render-timeline'.